Move GenerateRuntimeOptions() to pkg/cri/config
